+++ tests/asyncio/test_messages.py
@@ -153,7 +153,7 @@ class AssemblerTests(unittest.IsolatedAsyncioTestCase):
         self.assertEqual(message, "tea")
 
     async def test_get_resumes_reading(self):
-        """get resumes reading when queue goes below the high-water mark."""
+        """get resumes reading when queue goes below the low-water mark."""
         self.assembler.put(Frame(OP_TEXT, b"caf\xc3\xa9"))
         self.assembler.put(Frame(OP_TEXT, b"more caf\xc3\xa9"))
         self.assembler.put(Frame(OP_TEXT, b"water"))
@@ -169,6 +169,19 @@ class AssemblerTests(unittest.IsolatedAsyncioTestCase):
         # queue is below the low-water mark
         await self.assembler.get()
         self.resume.assert_called_once_with()
+
+    async def test_get_does_not_resume_reading(self):
+        """get does not resume reading when the low-water mark is unset."""
+        self.assembler.low = None
+
+        self.assembler.put(Frame(OP_TEXT, b"caf\xc3\xa9"))
+        self.assembler.put(Frame(OP_TEXT, b"more caf\xc3\xa9"))
+        self.assembler.put(Frame(OP_TEXT, b"water"))
+        await self.assembler.get()
+        await self.assembler.get()
+        await self.assembler.get()
+
+        self.resume.assert_not_called()
 
     async def test_cancel_get_before_first_frame(self):
         """get can be canceled safely before reading the first frame."""
@@ -302,7 +315,7 @@ class AssemblerTests(unittest.IsolatedAsyncioTestCase):
         self.assertEqual(fragments, ["t", "e", "a"])
 
     async def test_get_iter_resumes_reading(self):
-        """get_iter resumes reading when queue goes below the high-water mark."""
+        """get_iter resumes reading when queue goes below the low-water mark."""
         self.assembler.put(Frame(OP_BINARY, b"t", fin=False))
         self.assembler.put(Frame(OP_CONT, b"e", fin=False))
         self.assembler.put(Frame(OP_CONT, b"a"))
@@ -320,6 +333,20 @@ class AssemblerTests(unittest.IsolatedAsyncioTestCase):
         # queue is below the low-water mark
         await anext(iterator)
         self.resume.assert_called_once_with()
+
+    async def test_get_iter_does_not_resume_reading(self):
+        """get_iter does not resume reading when the low-water mark is unset."""
+        self.assembler.low = None
+
+        self.assembler.put(Frame(OP_BINARY, b"t", fin=False))
+        self.assembler.put(Frame(OP_CONT, b"e", fin=False))
+        self.assembler.put(Frame(OP_CONT, b"a"))
+        iterator = aiter(self.assembler.get_iter())
+        await anext(iterator)
+        await anext(iterator)
+        await anext(iterator)
+
+        self.resume.assert_not_called()
 
     async def test_cancel_get_iter_before_first_frame(self):
         """get_iter can be canceled safely before reading the first frame."""
@@ -366,6 +393,17 @@ class AssemblerTests(unittest.IsolatedAsyncioTestCase):
         # queue is above the high-water mark
         self.assembler.put(Frame(OP_CONT, b"a"))
         self.pause.assert_called_once_with()
+
+    async def test_put_does_not_pause_reading(self):
+        """put does not pause reading when the high-water mark is unset."""
+        self.assembler.high = None
+
+        self.assembler.put(Frame(OP_TEXT, b"caf\xc3\xa9"))
+        self.assembler.put(Frame(OP_BINARY, b"t", fin=False))
+        self.assembler.put(Frame(OP_CONT, b"e", fin=False))
+        self.assembler.put(Frame(OP_CONT, b"a"))
+
+        self.pause.assert_not_called()
 
     # Test termination
 
@@ -495,15 +533,28 @@ class AssemblerTests(unittest.IsolatedAsyncioTestCase):
     # Test setting limits
 
     async def test_set_high_water_mark(self):
-        """high sets the high-water mark."""
+        """high sets the high-water and low-water marks."""
         assembler = Assembler(high=10)
         self.assertEqual(assembler.high, 10)
+        self.assertEqual(assembler.low, 2)
 
-    async def test_set_high_and_low_water_mark(self):
-        """high sets the high-water mark and low-water mark."""
+    async def test_set_low_water_mark(self):
+        """low sets the low-water and high-water marks."""
+        assembler = Assembler(low=5)
+        self.assertEqual(assembler.low, 5)
+        self.assertEqual(assembler.high, 20)
+
+    async def test_set_high_and_low_water_marks(self):
+        """high and low set the high-water and low-water marks."""
         assembler = Assembler(high=10, low=5)
         self.assertEqual(assembler.high, 10)
         self.assertEqual(assembler.low, 5)
+
+    async def test_unset_high_and_low_water_marks(self):
+        """High-water and low-water marks are unset."""
+        assembler = Assembler()
+        self.assertEqual(assembler.high, None)
+        self.assertEqual(assembler.low, None)
 
     async def test_set_invalid_high_water_mark(self):
         """high must be a non-negative integer."""